RETAIN general modify 1. Remove unnecessary comments 2. model.train() for train, model.eval() for test 3. F.softmax -> nn.Softmax 4. add optimizer weight decay for l2 5. Amend rnn_options based on high performance params 6. Save trained model weights 7. Move code to retain_general_cel.py file
