+++ src/rnn_options.py
@@ -1,6 +1,16 @@
 class Options(object):
-    epochs = 1
+    epochs = 20
     batch_size = 64
-    learn_rate = 0.1
     hidden_dim = 256
-    lr = 0.0001
+    dim_input = 256
+    dim_emb = 256
+    dropout_input = 0.6
+    dropout_emb = 0.6
+    dim_alpha = 256
+    dim_beta = 256
+    dim_output = 2
+    dropout_context = 0.6
+    # l2 regularization
+    lr = 1e-4
+    weight_decay = 1e-5
+
